fix(pii): Do not panic on large regex (#1474)
In INC-202 we saw that project configs with large entries for
sensitiveFields are able to crash worker threads, because we compile the
list into a string with | and did not handle the CompiledTooBig error
returned by the regex crate.

This PR prevents the panic by making to_pii_config fallible. The
resulting failure mode is that calling scrub_data on events will fail,
and the event is dropped with outcome "internal".