A WIP commit that breaks compilation but updates envelope limiter
The envelope limiter needs to avoid calling redis again and needs to
remove items that need to be removed.  This does that but it is not
hooked up properly as the rate_limited_categories which are passed
through the processing are too simplistic and instead I want to pass
the enforcement through.  That still needs to be hooked up.