ref(system): Make the main runtime multi-threaded (#1498)
The main runtime hosts most services and will potentially become the
only runtime in the long run. In that scenario, it should be multi-
threaded. However, all expensive operations will likely also move to
spawn_blocking or the EnvelopeProcessor, so we may want to reduce the
thread count of the runtime.