fix(store): Catch invalid spans in light normalization (#1513)
Make sure that all instances of `ProcessingAction::InvalidTransaction`
are raised from light normalization. This was not the case for span
validation until now.

This PR moves the call to `TransactionProcessor` to light normalization
entirely. In a future PR, we might be able to remove the enum variant
`ProcessingAction::InvalidTransaction` and make `Annotated::apply`
infallible. That would make processors pure modifiers, instead of a mix
of modifiers and filters.

Co-authored-by: Oleksandr <olksdr@users.noreply.github.com>