feat(txquota): require tx_processed quota
If metrics have not yet been extracted a transaction event needs to
have DataCategory::TransactionProcessed quota available for ingestion.

Otherwise we are in the quota enforcement part of the relay event
processing pipeline and should check and count transaction indexing
quota.  Transaction processing quota however is only counted in the
metrics part and not in event ingestion.

INGEST-1581
INGEST-1653