fix(quotas): Make redis rate limiter work with quantity 0 (#1519)
For #1515, it is required to check for a required quota of another data
category without incrementing it. This PR updates the Redis LUA script
to support a rate limiting quantity of `0`, which checks for existing
rate limits without incrementing internal counters.

The rate limiter gains a new explicit branch to check whether the
quantity is `0`. 

Co-authored-by: Jan Michael Auer <mail@jauer.org>