ref(actix): Migrate Aggregator actor to new tokio runtime (#1508)
Migrate to new tokio runtime removing as many actix dependencies as possible.  
There are still few things kept temporary around to make the integration with old actix actors easier, but those will be removed in the followup PRs when the ProjectCache actor will be migrated to the new runtime as well.

The shutdown process and entire workflow of passing messages from one actor/service to another was kept identical to the old one, to keep the changes to minimum.