meta(vscode): Configure rust import preferences (#1522)
Configures Rust Analyzer to follow the preferred import style when
organizing imports, mostly during auto-import:

- Group imports by module path, emitting one line per module.
- Prefix imports from the local crate with `crate::` instead of using
  `super`.

Note that Rust Analyzer is not configured to rewrite existing imports,
and it will try to match the import style of the files being edited.

This also enables clippy as check command so that clippy warnings show
up while editing. This was likely already enabled in most local
development environments.