feat(quotas): Enforce rate limits on metrics buckets (#1515)
Apply rate limits to metrics buckets from the `transactions` namespace.

# Background 

Dynamic Sampling introduces a new type of quota,
`transactions_processed`, which defines how many transactions should be
"processed", i.e. whether or not we should extract metrics from them.
This quota is always >= the `transactions` quota, which defines how many
transactions should be stored & indexed.

The new data category should not only apply to incoming transaction
payloads (see https://github.com/getsentry/relay/pull/1517), but also to
metrics buckets, which may have been extracted by an upstream Relay.

Moreover, processing Relays must _count_ the number of transactions for
which metrics have been dropped, and create outcomes for them, in the
same way that we create accepted outcomes (see
https://github.com/getsentry/sentry/pull/39236).

# Data Flow

Processing Relays check rate limits against Redis. Currently the only
actor that has access to this rate limiter is the `EnvelopeProcessor`,
and we want to make use of its thread pool to make the blocking calls to
Redis. This makes it necessary to redirect metrics buckets to the
processor before sending them to the `EnvelopeManager` for publishing:

## Old Flow (still applies under certain conditions)

```
 ┌──────────┐    ┌────────────┐ ┌───────────────┐
 │Aggregator│    │ProjectCache│ │EnvelopeManager│
 └────┬─────┘    └─────┬──────┘ └──────┬────────┘
      │                │               │
      │  FlushBuckets  │               │
      ├───────────────►│               │
      │                │  SendMetrics  │
      │                ├──────────────►│
      │                │               │
```

## New Flow

```
 ┌──────────┐  ┌────────────┐       ┌─────────────────┐ ┌─────┐ ┌───────────────┐
 │Aggregator│  │ProjectCache│       │EnvelopeProcessor│ │Redis│ │EnvelopeManager│
 └────┬─────┘  └─────┬──────┘       └────────┬────────┘ └──┬──┘ └──────┬────────┘
      │              │                       │             │           │
      │ FlushBuckets │                       │             │           │
      ├─────────────►│                       │             │           │
      │              │ RateLimitFlushBuckets │             │           │
      │              ├──────────────────────►│             │           │
      │              │                       ├────────────►│           │
      │              │                       │             │           │
      │              │                       │◄────────────┤           │
      │              │                       │             │           │
      │              │                       │             │           │
      │              │                       │       SendMetrics       │
      │              │                       ├─────────────┬──────────►│
      │              │                       │             │           │
```

# Business Logic

1. ProjectCache received a `FlushBuckets` from the metrics aggregator.
2. Count the number of transactions that contributed to these buckets
(see below).
2. Check if cached rate limits apply.
3. If so, drop transaction-related buckets, generate outcomes, and send
the buckets to `EnvelopeManager`.
4. Otherwise, if processing mode is enabled, send buckets to the
`EnvelopeProcessor`.
5. The processor communicates the transaction count to redis and applies
the rate limit if reached.
6. The processor sends a message to the project cache to update the
cached rate limits.
7. Drop transaction-related buckets, generate outcomes, and send the
buckets to `EnvelopeManager`.

## Counting processed transactions

* For `d:transactions/duration@millisecond`, increment the counter in
redis by the length of the bucket's value. This is an accurate count of
the number of transactions that contributed to the bucket.
* For any other metric, do not increment the counter in redis, but still
enforce the rate limit if the quota is currently exceeded.

# Not in this PR
* Enforcement of rate limits on the fast path.
* Enforcements of rate limits on metrics buckets in the `sessions`
namespace.

Co-authored-by: Jan Michael Auer <mail@jauer.org>