fix(py): Respect the renormalize flag (#1548)
The `normalize` in the python package and C-ABI receives a structure
with config arguments for the various normalizers. The `renormalize`
flag is more central, as it is supposed to disable most of the
normalizers. This is used in Sentry after the event has passed
normalization to ensure that it is still structurally valid, but it may
contain fields now that are prohibited during ingestion.

In #1366 large parts of normalization were split into a dedicated
`light_normalize` function which did not honor the renormalize flag.
This PR restores correct behavior and disables the entire renormalize
call.

Co-authored-by: Jan Michael Auer <mail@jauer.org>