ref: Use wildcard re-exports where applicable (#1526)
In some cases, we export the entire public API of a module via the
containing parent module. The internal module structure is supposed to
be an implementation detail for code organization but does not matter to
the public API. An example for this is the Event protocol.

In most cases, we use wildcard exports `pub use module::*` for this. The
`relay-general` crate did not do this. With modern tooling and code
search, it is now equally easy to locate types. The benefit of using a
wildcard export is that we cannot forget to re-export types that would
be then missing from the public signature.

#skip-changelog