feat(quotas): Separate processing and indexing quotas (#1537)
Once metrics have been extracted, transaction events need to have the
new `DataCategory::TransactionIndexed` quota available to ingest and
store the payload.

The new indexing quota replaces the previous `DataCategory::Transaction`
for transaction payloads. Transaction quotas now cover all ingested
transactions that have metrics extracted. Transaction quota however is
only counted in the metrics part and not in event ingestion.

# Background

Rate limiting and quota enforcement has to apply to _legacy_ plans that
did not have processed transaction functionality, and to _new_ plans
that introduce them.

### Organization on Legacy Plan

There is a single quota for total transactions. We need to extract
metrics for indexed transactions so that there is backfilled history at
the time the customer switches to the new plan. This requires to extract
metrics for all accepted transactions.

### Organization on New Plan

This plan has a dedicated additional quota for indexed transactions.

- Validation and inbound data filters do not consume any quota. There
  are no metrics extracted for such dropped events.
- If there is transaction quota available, extract and ingest metrics
  for this transaction. Otherwise, drop the transaction event and any
  transaction metrics bucket.
- Dynamic sampling may drop (previously processed) transaction events.
  Extracted metrics and the transaction quotas are not impacted by this.
- If there is additional indexing quota available, store the sampled
  transaction event payload. Otherwise, drop the transaction event.

# Algorithm

**Flow:**

1. Check if metrics transaction is enabled.
   a. If disabled, rate limit the transaction item with the `transaction`
   data category and exit. 
   b. Otherwise, continue with step 2.
2. Check for transaction quota but do not increment it.
   a. If quota is exceeded, drop the item and all metrics.
   b. Otherwise, continue with step 3.
3. Check if metrics have been extracted for this transaction yet.
   a. If not, skip to step X
   b. Otherwise, keep the item. Metrics can be extracted now.
4. Metrics have been extracted. Check for indexing quota.
   a. If quota is exceeded, drop the item but keep metrics.
   b. Otherwise, keep the item. The transaction should be stored.

Transaction quota is incremented by the extracted metrics.

Co-authored-by: Joris Bayer <joris.bayer@sentry.io>