feat(dyn): Add support for decaying rules (#1544)
Decaying rules are a new set of rules to be applied in dynamic sampling.
They are temporary rules (i.e. rules with a start and end time) to alter
the sampling for some time. A simple use case is increasing the sampling
rate of a project for some time after a release.

Implementation-wise, they are basically normal rules. `timeRange` is the
new and optional parameter on every rule, and its existence determines
whether a rule is a decaying rule or not. `start` and `end`
subparameters are also optional, and they determine when the range
starts (inclusive) and ends (exclusive). When a limiter is not provided
relay doesn't restrict that end; for example, if `start` is provided but
`end` isn't, relay will apply the rule indefinitely after `start` time.
See the following rule example, taken from a unit test in this PR:

```json
{
    "condition":{
        "op":"and",
        "inner": [
                { "op" : "glob", "name": "releases", "value":["1.1.1", "1.1.2"]}
        ]
    },
    "sampleRate": 0.7,
    "type": "trace",
    "id": 1,
    "timeRange": {
        "start": "2022-10-10T10:10:10.101010Z",
        "end": "2022-10-20T20:20:20.202020Z"
    }
}
```

For an easy understanding of how the logic is intended to work, see the
unit test in
https://github.com/getsentry/relay/pull/1544/commits/d2d5f6a5e7c00eaca2546ca25f27ffd6873e77f8.