fix(quotas): Correct reporting of rate limits for transactions (#1555)
With #1537 Relay started to under-report outcomes for rate limited
transactions. The root cause is that `Enforcement` was not updated to
emit outcomes for metrics.

This reveals that enforcements for metrics should be recorded only if
metrics have not been extracted from a transaction item yet. In this
case, metrics can still be considered part of the transaction item and
as such an outcome is needed. In contrast, once metrics have been
extracted, responsibility for recording outcomes is moved to the
recipient of metrics.