ref(event): Add correct limit and validation to distribution name
Even though we had the `max_chars` limit set to 200 on the `dist` we
never validated that length while trying to normalize the event.

This change adds the correct limit (64 chars) which is set on the Sentry
site (field in the DB is varchar(64)) and also adds the validation with
proper error generation if the value longer than the length limit.