fix(tests): Reduce rate limiting test flakyness (#1568)
Rate limiting tests use real time. With small rate limiting windows,
chances are higher that the window changes between two event
submissions. To counter this, we've so far increased the window of the
test quotas to one day, which makes it extremely unlikely.

A proper solution for this problem would be to set explicit timestamps
and move such tests to unit-test level.