fix(protocol): Disallow `-` in measurement and breakdown names (#1571)
Names of custom measurements are converted into MRIs without any
validation. This means that in metrics extraction, we are able to
construct MRIs that are less strict than what the metrics protocol
demands. Subsequent attempts to parse the MRI string into a
`MetricsResourceIdentifier` object will fail.

This popped up in rate limiting, where we have to parse the MRI to know
what namespace and name we are dealing with. Because of this bug, custom
measurements with invalid names are currently not rate limited at all.

Fixes
[POP-RELAY-2N8](https://sentry.my.sentry.io/organizations/sentry/issues/315035).