feat(system): Log service back pressure as metric (#1583)
This adds central back pressure monitoring via statsd metrics to all
services that use the new tokio-based service framework. Channels track
the number of pending messages and emit a gauge metric with that number.

The metric is `service.back_pressure` and tagged with a `service` tag to
identify the specific service. The metric is debounced once per second.
Between these intervals, the back pressure number is not checked, so
this metric is not suitable to detect very short-lived spikes.
Additionally, the current implementation requires that the service makes
progress by polling `recv`.

The purpose of this metric is to detect lasting increases in backlogs
that indicate bottlenecks or degraded system behavior.