feat(sampling): Allow boosting individual transactions (#1581)
## Background

We want to boost sample rates for key transactions. Previously, a
combination of trace and event sampling rules would lead to a
multiplication of sample rates. For example, a base rate of `0.1` and a
boost rule of `0.5` would lead to an effective sampling rate of

```
0.1 * 0.5 == 0.05
```

The reason for this is that trace rules use a different random number
seed as transaction rules (trace-ID vs. event-ID), so we effectively
rolled the dice twice.

## Solution

* As soon as a rule matches, make a sampling decision (roll the dice)
and exit (to prevent double sampling).
* When confronted with two different rule sets (trace rules and event
rules), let event rules take precedence. This allows us to define boost
rules for key transactions as well as reducing rules for health checks.

## Drawbacks

With the new behavior, it is not possible to make a trace rule (e.g. a
boost rule for recent releases) take precedence over an active
transaction rule.