ref(system): Introduce a generic recipient type (#1622)
`Recipient` is a type similar to `Addr`, but it is independent of
interfaces. This makes it suitable for inversion of control, where a
service would like to send messages to other services without knowing
their interfaces.

This may not be the final implementation of the IoC pattern. This
initial implementation only supports a single message and not entire
interfaces (or extending interfaces). It is modeled after `actix::Addr`.
In the future, Recipient may be refactored to support entire interfaces.