feat(system): Add broadcast behavior for debounced services (#1619)
The `ProjectCache` and `RelayCache` actors both follow a common pattern:
They maintain a map of cached objects that they populate asynchronously.
There is a message in their public interface to retrieve data from this
cache.

Since the `RelayCache` was ported in #1485, it keeps a list of senders
for each waiting caller of that message. Once the cache entry has been
resolved, it loops through all senders and passed a clone of the value
on. This has the downside that it doesn't only require resources on the
requesting side, but also in the cache actor itself.

A better approach is to break this into a two-step process:

1.  The service creates a shared channel that will be populated once
2.  Recipients attach to the shared channel and wait for the response

The recipient therefore first has to await the receive end of the shared
channel, and then await the value from that channel. Of course, if the
value is already in the cache, this can be skipped, and the value can be
sent directly to the recipient.

# Implementation Details

This PR implements this with two nested `oneshot` channels. The outer
channel resolves an enum, which is either the result value, or a shared
inner channel that resolves the value.

We implement this as service response behavior, so that the building
blocks for this (senders, channels, ...) can be reused between multiple
services. This can be used for some of the messages, while other
messages have regular async behavior. The response future type is
opaque, so users of such a service do not notice the two-step behavior;
they simply resolve the final value.

This PR updates the `RelayCacheService` to use this new response
behavior. Internally, the service still needs to hold two maps:

- A map for the actual raw data that it pulls responses from. This data
  can be different or more elaborate than what's exposed via public
  messages.
- A map for open channels. Entries in this map can be removed as soon as
  the cache value is resolved, because after that the service can
  short-circuit and respond with the value directly.