fix(server): Forward-compatible AttachmentType (#1638)
As we did in https://github.com/getsentry/relay/pull/1246 for
`ItemType`, add an `Unknown(String)` variant to `AttachmentType`. When
`accept_unknown_items` is set to false, items with this attachment type
will be dropped.

This will allow outdated external Relays to forward new attachment types
instead of dropping the entire envelope.

This defect was discovered in https://github.com/getsentry/rfcs/pull/33,
which introduces a new attachment type.