fix(server): Drop transactions outside of the metrics time range (#1663)
Relay used to accept transactions back-dated up to 30 days into the
past, but transaction metrics are restricted to 5 days, a restriction
which our storage relies on. With this PR, metric extraction now
validates the transaction timestamps and does not extract metrics if the
timestamp is outside the allowed range for metrics.

Additionally, Relay also needs to drop the transaction event for
consistency between metrics and events. Otherwise, a transaction may be
indexed that is not included in the "total" metrics dataset. This would
lead to inconsistency between the datasets and violate an invariant that
metrics contain a superset of indexed transactions.