ref(ffi,cabi): Replace outdated failure with anyhow crate (#1669)
Failure crate is outdated and not supported anymore. It's time to switch
to modern alternative.
Instead of `failure::Error` switching to `anyhow::Error` which is a
wrapper around a dynamic error type.

`Error` works a lot like `Box<dyn std::error::Error>`, but with these
differences:

* Requires that the error is Send, Sync, and 'static.
* Guarantees that a backtrace is available, even if the underlying error
type does not provide one.
* Is represented as a narrow pointer — exactly one word in size instead
of two.