fix(sampling): Sample transactions without DSC (#1667)
Dynamic sampling relies on the Dynamic Sampling Context (DSC). The
envelope header contains information about the trace, the origin
project, and attributes for trace sampling.

Originally, when the DSC is not specified in an envelope, dynamic
sampling did not run on the envelope. Transaction events in this
envelope are effectively sampled at 100% percent. However, there are
more cases where the DSC is not available or not valid:

- Old or incompatible SDKs do not send a DSC, but the transaction needs
  to be sampled at least with the uniform sample rate.
- The DSN public key in the DSC could point to a nonexistent project or
  a project from another organization.
- There could be an error resolving the project state for the specified
  project.

This PR makes the following changes:

- Uses the event to generate an ad-hoc DSC for trace sampling if there's
  no DSC in the event or the sampling project state cannot be retrieved
  for any reason.
- Ensures that only rules from the same organization are used for trace
  sampling, falling back to the local project.