Publish Docker images on DockerHub via Github Action
This PR simply adds the docker login information (to be stored in a secret) so that the self-hosted e2e test action can push to DockerHub. For now the action pushes to a different tag, as not to conflict with Google Cloud Build pushing to the same repo at the same time. This will be changed once we are confident this action is stable and we can remove Google Cloud Build outright.