Adding support for transaction (re)naming rules
These changes introduce the support for naming rules, which currently
can be applied to only `url` transaction source:

* relay always applys first matching rule
* rules applied in light normalization, before extraxting any
  information from the event
* original transaction name is preserved in meta, and also remark added
  with the rule, which was applied
* `transaction_info.source` is changed to `sanitized` after rule is
  applied

This PR also contains changes to the current implementation of `Glob`,
addint the ability to apply the matching rule to replace parts of the
matching pattern - which maybe could be done separatelly.

Also few existing structs were re-used / re-purposed to support new rule
format, which also up for the discussions.