revert(profiling): Remove concurrent profiling (#1697)
This piece of code was duplicating profiles and trimming them to the
right duration for each transaction they were associated with.

As we want to be able to only accept profiles associated with a
transaction, we have to send the profile in the same envelope as
transaction so dynamic sampling works properly. We can't start multiple
profilers at the time (Android API we're using doesn't allow that) or we
can't do the processing the server was doing to get one profile and trim
on the client. We will keep trimming `sample` and `cocoa` formats.

Therefore, we decided to remove this concurrent profiling support.

Co-authored-by: Jan Michael Auer <mail@jauer.org>