fix(crash): Use main SDK to send crash reports (#1707)
Relay captures crashes with the Sentry Native SDK and its breakpad
backend. By default, this uses curl as transport. In our Docker build,
we dynamically link with the installed development library. Since the
build runs on CentOS 7, this is an old library lacking many security
features and patches. Even though the production container has a more
recent libcurl installed, the Native SDK uploader fails the SSL
handshake with sentry.io.

This PR removes the dependency on curl entirely and replaces it with a
custom transport function. We parse the envelope and send it using the
main Sentry SDK.