ref(actix): Migrate RedisProjectSource from actor to utility class (#1714)
Instead of using the `actix::Actor` and instead of migrating this to our
new `Service` this change introduces the simple utility class
`RedisProjectSource` which just provides only one helper method to fetch
project config.

The call to this method is wrapped into `tokio::task::spawn_blocking`
which should be executed on the separated thread pool. And the rest of
the functionality stays practically the same.