ci: Skip redundant self-hosted E2E on library release (#1755)
Library release builds are configured to skip docker builds since those
are not actually required. When self-hosted E2E tests were added, it was
missed to exclude them from running on library releases. They are
actually redundant, since they do not test the new build of the library
but just the Relay binary.