ref(statsd): Add metric name as a tag for Sentry errors (#1797)
From time to time, we still get errors from relay dropping metrics due
to the channel being full. This commit attempts to add
a tag to that Sentry issue, to identify more easily what metrics may be
worsening the bottleneck. Next steps TBD, based on conclusions from the
errors. This code is likely to be temporary and be removed after we
figure out what and how to deal with the errors.