ref(system): Add helper to start services in runtimes (#1458)
When creating Relay's services network, we commonly spawn groups of
services into dedicated runtimes. There is a repeated pattern of
creating a runtime, entering it, starting the service, dropping the
runtime guard.

Which runtime each service runs in depends on the runtime guard that was
last activated before. This is not as clear and safe in the code as it
could be. For that reason, this PR introduces a `spawn_in` that takes
care of all the above steps and returns the address.

Two services needed refactoring as a result: The `OutcomeProducer` and
`ProjectCache` implementations. Both spawned sub-services in their
constructors, which escaped the scope created by `spawn_in`. To ensure
that the sub-services run within the same runtime, all spawning of
services has moved into the `spawn_handler` function. Note that
`ProjectCache` now assumes a similar pattern to the `UpstreamRelay` with
an internal broker and message channel.