feat(profiling): Add or remove the profile ID on the transaction's profiling context (#1801)
We added a way to set the profile ID in a profile context on the
transaction. Now, we need to set the `profile_id` or remove it from the
context if it's already set by the SDK in case the profile succeeds or
failed to be processed in order to have accurate transaction to profile
link. We don't want a transaction set with a profile ID that doesn't
exist and we don't want a transaction without a profile ID that do
exist.

This PR will process the profiles after the transaction is processed and
then take the appropriate action and modify the profile context on the
transaction.