feat: Add client hints for requestmeta (#1802)
Second PR about supporting client hints. #1183 
previous PR: #1752 

This PR adds support for client hints on the requestmeta struct, which
means that the client hints can be sent as part of an event json as an
optional field. This means that more information will be extracted from
the client hints rather than the user agent string, as relay will first
attempt to check the client hints and only falling back to the user
agent string if no client hints are found.

---------

Co-authored-by: Joris Bayer <joris.bayer@sentry.io>