ref: Make the pattern regex compilation lazy (#1829)
As a followup for #1819 (based on the comment
https://github.com/getsentry/relay/pull/1819#discussion_r1097515998)
this PR converts the `Pattern` to `LazyPattern` and delays the `Regex`
compilation to the first usage. But we do keep the incoming raw string
pattern around.

If the regex cannot be compiled, we avoid creating the redact rules,
which keeps the similar behaviour, since before this change we just
error out on the deserialization.