bug(replays): Add default user object and fill its ip-address (#1805)
If the User object is null we currently skip the ip normalization step.
Now (if we have an ip-address) we will add a default User if one does
not exist and add the ip-address to that User. This fixes a bug where a
large number of replays were being ingested without an ip-address.

---

Edit: out of date.  Left for historical purposes.

I also observed this functionality:


https://github.com/getsentry/relay/blob/01514478bb1614124c360407b599c525d1371c38/relay-general/src/store/normalize.rs#L593-L650

Obviously my additions are much less than what this function
encompasses. Some of this seems related to back-end SDKs which may be
outside of scope. I did notice lines 617-620 seem to handle the case
where the user's ip was submitted as `auto`. Is this something the JS
SDKs do?

---------

Co-authored-by: Joris Bayer <joris.bayer@sentry.io>