build(log): Upgrade log dependencies (#1847)
Updates `env_logger`. For human readable output we have used
`pretty_env_logger`, which is no longer compatible with the latest
`env_logger` version. This PR vendors the relevant bits of the
implementation.

Previously, the `relay-log` crate hard-coded a list of all workspace
members to set log levels based on the configuration. This has been
refactored:

- The list of workspace members is now created in a build script.
- The default level filter sets every workspace member to TRACE.
- There is a global `max_level` filter based on the overall log level
  config.