ref(server): Move ProjectConfig to separate crate (#1858)
In order to expose a function in `relay-cabi` that validates project
configs, we have to move `ProjectConfig` out of `relay-server` to
prevent that crate becoming a dependency. We briefly considered moving
it to `relay-general`, but decided that that crate is already too large
and unspecific.

Instead, move `ProjectConfig` to a new crate `relay-dynamic-config`. The
name foreshadows future additions such as org-level config or even
dynamic global config passed down from sentry.

This PR does not make any functional changes.