ref: Buffer envelopes for broken project states (#1856)
Check request only for the valid project configs, otherwise we reject
incoming events immediately with `invalid data (project_state)`, since
there is nothing to check against.

With this change we will still keep buffering even if the resolved
project state is invalid. After the state is updated, the pending
envelopes will be processed again.

We also will keep trying to fetch the new state for the project, if it was invalid
to stay proactive and get fresh usable state as soon as possible.