feat(metrics): Tag the sample decision on count_per_root_project (#1870)
Adds a binary `decision` tag to `c:transactions/count_per_root_project`.
This allows us to query the effective sample rate on projects.

Background: Dynamic sampling is based on a target sample rate, called
the organization's _"fidelity"_. On top of that, Sentry computes a set
of _priorities_ (formerly called "biases") to increase visibility in
under-represented areas. For example, Sentry boosts releases during
adoption or small projects.

The priorities currently apply on top of the fidelity rate and increase
the sample rate for a subset of transactions. Overall, this increases
the effective sample rate applied for a project. The extent of this is
not deterministic, as it depends on the match rate of sample rules.

The final goal is to adjust the priorities and base sample rate in such
a way that the effective sample rate matches the target fidelity.