feat(threads): Add state as an optional key in the threads interface (#1896)
Thread state is an optional string currently sent by the Android SDK on
the threads interface but isn't extracted. We want to extract and store
it on the event json. Currently, we capture the JVM
[states](https://docs.oracle.com/javase/7/docs/api/java/lang/Thread.State.html),
but in the ANRv2 implementation, we expect to capture to states the
Android VM
[provides](https://cs.android.com/android/platform/superproject/+/master:art/runtime/thread_state.h).
Note, these values are for Android only, other SDKs could send other
values.

We only need to save thread state in nodestore, we don't anticipate
running any queries against this field.


[Here](https://gist.github.com/shruthilayaj/2151eac05009466f4549c5f2ad48a11f)
is an example event payload with thread states.

closes https://github.com/getsentry/sentry/issues/45196